add replaceMany addon
