+++ browser/mocha.js
@@ -12,6 +12,7 @@ import * as config from './config.js';
 import './mocha/fixture.js';
 import './mocha/stub.js';
 import './mocha/replace.js';
+import './mocha/replace-many.js';
 import { applyExtensions } from './mocha/extend.js';
 
 // Mocha global helpers, broken out by testing method.
@@ -86,4 +87,3 @@ function _setupMocha(ui, key, alternate) {
   mochaOptions.ui = ui;
   mocha.setup(mochaOptions);  // Note that the reporter is configured in run.js.
 }
-
