+++ browser/mocha/replace.js
@@ -1,4 +1,5 @@
 import { extendInterfaces } from './extend';
+import { stubInstanceTemplate, tagReplacer } from './helper';
 
 /**
  * replace
@@ -18,46 +19,8 @@ extendInterfaces('replace', function(context, teardown) {
   return function replace(oldTagName) {
     return {
       with: function(tagName) {
-        // Keep a reference to the original `Polymer.Base.instanceTemplate`
-        // implementation for later:
-        var originalInstanceTemplate = Polymer.Base.instanceTemplate;
-
-        // Use Sinon to stub `Polymer.Base.instanceTemplate`:
-        sinon.stub(Polymer.Base, 'instanceTemplate', function(template) {
-          // The DOM to replace is the result of calling the "original"
-          // `instanceTemplate` implementation:
-          var dom = originalInstanceTemplate.apply(this, arguments);
-
-          // The nodes to replace are queried from the DOM chunk:
-          var nodes = Array.prototype.slice.call(dom.querySelectorAll(oldTagName));
-
-          // For all of the nodes we want to place...
-          nodes.forEach(function(node) {
-
-            // Create a replacement:
-            var replacement = document.createElement(tagName);
-
-            // For all attributes in the original node..
-            for (var index = 0; index < node.attributes.length; ++index) {
-              // Set that attribute on the replacement:
-              replacement.setAttribute(
-                node.attributes[index].name, node.attributes[index].value);
-            }
-
-            // Replace the original node with the replacement node:
-            node.parentNode.replaceChild(replacement, node);
-          });
-
-          return dom;
-        });
-
-        // After each test...
-        teardown(function() {
-          // Restore the stubbed version of `Polymer.Base.instanceTemplate`:
-          if (Polymer.Base.instanceTemplate.isSinonProxy) {
-            Polymer.Base.instanceTemplate.restore();
-          }
-        });
+        var decorator = tagReplacer(oldTagName, tagName);
+        return stubInstanceTemplate(decorator, teardown);
       }
     };
   };
